requirements.txt
alabaster==0.7.12
amqp==2.3.2
argh==0.26.2
asn1crypto==0.24.0
Babel==2.6.0
billiard==3.5.0.5
celery==4.2.1
certifi==2018.10.15
cffi==1.11.5
chardet==3.0.4
CommonMark==0.5.4
cryptography==2.4.2
Django==2.1.3
django-guardian==1.4.9
django-widget-tweaks==1.4.3
djangorestframework==3.9.0
docutils==0.14
idna==2.7
imagesize==1.1.0
Jinja2==2.10
kombu==4.2.1
livereload==2.6.0
MarkupSafe==1.1.0
packaging==18.0
pathtools==0.1.2
port-for==0.3.1
pycparser==2.19
Pygments==2.2.0
PyMySQL==0.9.2
pyparsing==2.3.0
pytz==2018.7
PyYAML==3.13
recommonmark==0.4.0
requests==2.20.1
six==1.11.0
snowballstemmer==1.2.1
Sphinx==1.8.2
sphinx-autobuild==0.7.1
sphinx-rtd-theme==0.4.2
sphinxcontrib-websupport==1.1.0
tornado==5.1.1
urllib3==1.24.1
vine==1.1.4
watchdog==0.9.0
